autobuild: add autobuild scripts
The autobuild scripts provide an unified way
to build agl apps & services; for the agl
modules missing these scripts, a workaround is
in place to warn of this scenario and allow
for build completion.
This commit add these missing scripts for
genivi-navi-yelp-client (aka poiapp)
The original scripts are slimmed down and adapted
to the existing cmake files without modifying them.

- autobuild/agl/autobuild: allows aglwgt_package
  task to complete, eliminating the verbose
  warnings.
- autobuild/linux/autobuild: allows to build
  the wgt file natively on linux using the agl
  sdk.

Signed-off-by: Raquel Medina <raquel.medina@konsulko.com>